+++ libera_utils/io/smart_open.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:0ee14423bfec30f93209b6098c2a308a7019af18399578edec4be88800368b3b
-size 9930
+oid sha256:27f5efdf4edc7e5d47c5daa66564ce26d8caa950af9f19fa12f4e74c0bca5a40
+size 9909