+++ tests/plugins/aws_fixtures.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:97ea8c43182898436db675ceacb09814a5e0d2824bdee3af9b6c133a35a15931
-size 7807
+oid sha256:ed5a7d6dd97b221683bc86d282b2a22630989fe35459f97d67222b9e504f4c19
+size 8136