+++ tests/unit/test_cli.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1ba95d0cd781b4ce8096bd5d78e964f0b420522907e32aee3ff4a89f831bb97a
-size 1971
+oid sha256:fe62390d83330fd2d5f7e9d2d9342a56062c7a7ac9dffae922ff7810fda2dcbb
+size 4452