+++ tests/unit/test_io/test_smart_open.py
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:dff0294e7144024ede6a2d2a863ef38d17a1a4479b5858af312fabeaed403529
-size 16000
+oid sha256:9cc7e235b5a6abacb74fec9659f71f8742784a9e73afdf24cb5ffa20df59d5f4
+size 16214