+++ doc/source/changelog.md
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3d93053ff03aca8e4f25c82855d859a4d4100825510b23ce767bb50811c54d82
-size 5161
+oid sha256:a15437fd321ae14b05a0f638ec22df81137ae3a168cbb2499e395757a9a907da
+size 5849